[Migrate to Symfony] Migrate Ingredient Entity
