+++ src/AppBundle/Entity/IngredientRepository.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace AppBundle\Entity;
+
+use Doctrine\ORM\EntityRepository;
+
+/**
+ * Class IngredientRepository.
+ */
+class IngredientRepository extends EntityRepository
+{
+}
