+++ src/AppBundle/Entity/Ingredients.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Lunches\Model;
+namespace AppBundle\Entity;
 
 use Doctrine\Common\Collections\ArrayCollection;
